clean up schema normalization
