+++ libs/skills/swarm/executor.test.ts
@@ -32,7 +32,7 @@ describe("callTask", () => {
     expect(result).toBe("Result for: do something");
   });
 
-  it("forwards response_schema when provided (normalized)", async () => {
+  it("forwards response_schema when provided", async () => {
     const schema = { type: "object", properties: { x: { type: "number" } } };
     await callTask({
       description: "structured",
@@ -47,11 +47,7 @@ describe("callTask", () => {
     expect(taskFn.task).toHaveBeenCalledWith({
       description: "structured",
       subagent_type: "analyst",
-      response_schema: {
-        type: "object",
-        additionalProperties: false,
-        properties: { x: { type: "number" } },
-      },
+      response_schema: schema,
     });
   });
 
@@ -171,7 +167,6 @@ describe("dispatch", () => {
     );
     expect(taskFn.mock.calls[1][0].response_schema).toEqual({
       type: "object",
-      additionalProperties: false,
     });
   });
